Fix faulted zones clearing iteration logic (#77)
As provided by warthog618 in https://github.com/nutechsoftware/alarmdecoder/issues/54

The restoration based on message.ready is buggy as it iterates over a list it is deleting from.  which results in only half the faulted zones being cleared.  The fix is to iterate over a copy of the list.